reorganization of the project structure
