+++ components/daikin_rotex_can/translations/de.py
@@ -1,0 +1,105 @@
+"""
+    This file is used to store translations for the Daikin Rotex CAN component.
+    Language: German
+"""
+
+translations_de = {
+    "heating": "Heizen",
+    "cooling": "Kühlen",
+    "standby": "Standby",
+    "hot_water_production": "Warmwasserbereitung",
+    "defrosting": "Abtauen",
+    "optimized_defrosting": "Optimiertes Abtauen",
+    "temperature_antifreeze": "T-Frostschutz",
+    "off": "Aus",
+    "on": "An",
+    "night_only": "Nur bei Nacht",							  
+    "monday": "Montag",
+    "tuesday": "Dienstag",
+    "wednesday": "Mittwoch",
+    "thursday": "Donnerstag",
+    "friday": "Freitag",
+    "saturday": "Samstag",
+    "sunday": "Sonntag",
+    "mo_to_su": "Montag bis Sonntag",
+    "lowering": "Absenken",
+    "summer": "Sommer",
+    "automatic_1": "Automatik 1",
+    "automatic_2": "Automatik 2",
+    "weather_dependent": "Witterungsgeführt",		
+    "standby_mode": "Standby",
+    "heating_mode": "Heizen",
+    "cooling_mode": "Kühlen",
+    "defrosting_mode": "Abtauen",
+    "hot_water_production_mode": "Warmwasserbereitung",
+    "sg_mode_1": "SG Modus 1",
+    "sg_mode_2": "SG Modus 2",
+    "fixed": "Fest",
+    "no_additional_heat_generator": "Kein zusätzlicher Wärmeerzeuger",
+    "optional_backup_heater": "Optionaler Backup-Heater",
+    "wez_for_hot_water_and_heating": "WEZ für WW und HZ",
+    "wez1_for_hot_water_wez2_for_heating": "WEZ1 für WW - WEZ2 für HZ",
+    "err_0": "Kein Fehler",
+    "err_E9001": "E9001 Rücklauffühler",
+    "err_E9002": "E9002 Vorlauffühler",
+    "err_E9003": "E9003 Frostschutzfunktion",
+    "err_E9004": "E9004 Durchfluss",
+    "err_E9005": "E9005 Vorlauftemperaturfühler",
+    "err_E9006": "E9006 Vorlauftemperaturfühler",
+    "err_E9007": "E9007 Platine IG defekt",
+    "err_E9008": "E9008 Kältemitteltemperatur außerhalb des Bereiches",
+    "err_E9009": "E9009 STB Fehler",
+    "err_E9010": "E9010 STB Fehler",
+    "err_E9011": "E9011 Fehler Flowsensor",
+    "err_E9012": "E9012 Fehler Vorlauffühler",
+    "err_E9013": "E9013 Platine AG defekt",
+    "err_E9014": "E9014 P-Kältemittel hoch",
+    "err_E9015": "E9015 P-Kältemittel niedrig",
+    "err_E9016": "E9016 Lastschutz Verdichter",
+    "err_E9017": "E9017 Ventilator blockiert",
+    "err_E9018": "E9018 Expansionsventil",
+    "err_E9019": "E9019 Warmwassertemperatur > 85°C",
+    "err_E9020": "E9020 T-Verdampfer hoch",
+    "err_E9021": "E9021 HPS-System",
+    "err_E9022": "E9022 Fehler AT-Fühler",
+    "err_E9023": "E9023 Fehler WW-Fühler",
+    "err_E9024": "E9024 Drucksensor",
+    "err_E9025": "E9025 Fehler Rücklauffühler",
+    "err_E9026": "E9026 Drucksensor",
+    "err_E9027": "E9027 Aircoil-Fühler Defrost",
+    "err_E9028": "E9028 Aircoil-Fühler temp",
+    "err_E9029": "E9029 Fehler Kältefühler AG",
+    "err_E9030": "E9030 Defekt elektrisch",
+    "err_E9031": "E9031 Defekt elektrisch",
+    "err_E9032": "E9032 Defekt elektrisch",
+    "err_E9033": "E9033 Defekt elektrisch",
+    "err_E9034": "E9034 Defekt elektrisch",
+    "err_E9035": "E9035 Platine AG defekt",
+    "err_E9036": "E9036 Defekt elektrisch",
+    "err_E9037": "E9037 Einstellung Leistung",
+    "err_E9038": "E9038 Kältemittel Leck",
+    "err_E9039": "E9039 Unter/Überspannung",
+    "err_E9041": "E9041 Übertragungsfehler",
+    "err_E9042": "E9042 Übertragungsfehler",
+    "err_E9043": "E9043 Übertragungsfehler",
+    "err_E9044": "E9044 Übertragungsfehler",
+    "err_E75": "E75 Fehler Außentemperaturfühler",
+    "err_E76": "E76 Fehler Speichertemperaturfühler",
+    "err_E81": "E81 Kommunikationsfehler Rocon",
+    "err_E88": "E88 Kommunikationsfehler Rocon Handbuch",
+    "err_E91": "E91 Kommunikationsfehler Rocon Handbuch",
+    "err_E128": "E128 Fehler Rücklauftemperaturfühler",
+    "err_E129": "E129 Fehler Drucksensor",
+    "err_E198": "E198 Durchflussmessung nicht plausibel",
+    "err_E200": "E200 Kommunikationsfehler",
+    "err_E8005": "E8005 Wasserdruck in Heizungsanlage zu gering",
+    "err_E8100": "E8100 Kommunikation",
+    "err_E9000": "E9000 Interne vorübergehende Meldung",
+    "err_W8006": "W8006 Warnung Druckverlust",
+    "err_W8007": "W8007 Wasserdruck in Anlage zu hoch",
+    "defect": "Defekt",
+    "sgn_normal_mode": "SGN - Normaler Modus",
+    "sg1_hot_water_and_heating_off": "SG1 - WW & HZ ausgeschalten",
+    "sg2_hot_water_and_heating_plus_5c": "SG2 - WW & HZ + 5°C",
+    "sg3_hot_water_70c": "SG3 - WW 70°C"
+}
