+++ components/daikin_rotex_can/translations/en.py
@@ -1,0 +1,105 @@
+"""
+    This file is used to store translations for the Daikin Rotex CAN component.
+    Language: English
+"""
+
+translations_en = {
+    "heating": "Heating",
+    "cooling": "Cooling",
+    "standby": "Standby",
+    "hot_water_production": "Hot Water Production",
+    "defrosting": "Defrosting",
+    "optimized_defrosting": "Optimized Defrosting",
+    "temperature_antifreeze": "Antifreeze Temperature",
+    "off": "Off",
+    "on": "On",
+    "night_only": "Night Only",						   
+    "monday": "Monday",
+    "tuesday": "Tuesday",
+    "wednesday": "Wednesday",
+    "thursday": "Thursday",
+    "friday": "Friday",
+    "saturday": "Saturday",
+    "sunday": "Sunday",
+    "mo_to_su": "Monday to Sunday",
+    "lowering": "Lowering",
+    "summer": "Summer",
+    "automatic_1": "Automatic 1",
+    "automatic_2": "Automatic 2",
+    "weather_dependent": "Weather Dependent",		
+    "standby_mode": "Standby",
+    "heating_mode": "Heating",
+    "cooling_mode": "Cooling",
+    "defrosting_mode": "Defrosting",
+    "hot_water_production_mode": "Hot Water Production",
+    "sg_mode_1": "SG Mode 1",
+    "sg_mode_2": "SG Mode 2",
+    "fixed": "Fixed",
+    "no_additional_heat_generator": "No Additional Heat Generator",
+    "optional_backup_heater": "Optional Backup Heater",
+    "wez_for_hot_water_and_heating": "WEZ for Hot Water and Heating",
+    "wez1_for_hot_water_wez2_for_heating": "WEZ1 for Hot Water - WEZ2 for Heating",
+    "err_0": "No Error",
+    "err_E9001": "E9001 Return Sensor",
+    "err_E9002": "E9002 Supply Sensor",
+    "err_E9003": "E9003 Frost Protection Function",
+    "err_E9004": "E9004 Flow",
+    "err_E9005": "E9005 Supply Temperature Sensor",
+    "err_E9006": "E9006 Supply Temperature Sensor",
+    "err_E9007": "E9007 IG Board Defective",
+    "err_E9008": "E9008 Refrigerant Temperature Out of Range",
+    "err_E9009": "E9009 STB Error",
+    "err_E9010": "E9010 STB Error",
+    "err_E9011": "E9011 Flow Sensor Error",
+    "err_E9012": "E9012 Supply Sensor Error",
+    "err_E9013": "E9013 AG Board Defective",
+    "err_E9014": "E9014 High Refrigerant Pressure",
+    "err_E9015": "E9015 Low Refrigerant Pressure",
+    "err_E9016": "E9016 Compressor Overload Protection",
+    "err_E9017": "E9017 Fan Blocked",
+    "err_E9018": "E9018 Expansion Valve",
+    "err_E9019": "E9019 Hot Water Temperature > 85°C",
+    "err_E9020": "E9020 High Evaporator Temperature",
+    "err_E9021": "E9021 HPS System",
+    "err_E9022": "E9022 Ambient Temperature Sensor Error",
+    "err_E9023": "E9023 Hot Water Sensor Error",
+    "err_E9024": "E9024 Pressure Sensor",
+    "err_E9025": "E9025 Return Sensor Error",
+    "err_E9026": "E9026 Pressure Sensor",
+    "err_E9027": "E9027 Air Coil Defrost Sensor",
+    "err_E9028": "E9028 Air Coil Temperature Sensor",
+    "err_E9029": "E9029 Cooling Sensor Error AG",
+    "err_E9030": "E9030 Electrical Defect",
+    "err_E9031": "E9031 Electrical Defect",
+    "err_E9032": "E9032 Electrical Defect",
+    "err_E9033": "E9033 Electrical Defect",
+    "err_E9034": "E9034 Electrical Defect",
+    "err_E9035": "E9035 AG Board Defective",
+    "err_E9036": "E9036 Electrical Defect",
+    "err_E9037": "E9037 Power Setting",
+    "err_E9038": "E9038 Refrigerant Leak",
+    "err_E9039": "E9039 Under/Over Voltage",
+    "err_E9041": "E9041 Communication Error",
+    "err_E9042": "E9042 Communication Error",
+    "err_E9043": "E9043 Communication Error",
+    "err_E9044": "E9044 Communication Error",
+    "err_E75": "E75 Ambient Temperature Sensor Error",
+    "err_E76": "E76 Storage Temperature Sensor Error",
+    "err_E81": "E81 Communication Error Rocon",
+    "err_E88": "E88 Communication Error Rocon Manual",
+    "err_E91": "E91 Communication Error Rocon Manual",
+    "err_E128": "E128 Return Temperature Sensor Error",
+    "err_E129": "E129 Pressure Sensor Error",
+    "err_E198": "E198 Flow Measurement Not Plausible",
+    "err_E200": "E200 Communication Error",
+    "err_E8005": "E8005 Water Pressure in Heating System Too Low",
+    "err_E8100": "E8100 Communication",
+    "err_E9000": "E9000 Temporary Internal Message",
+    "err_W8006": "W8006 Pressure Loss Warning",
+    "err_W8007": "W8007 Water Pressure in System Too High",
+    "defect": "Defect",
+    "sgn_normal_mode": "SGN - Normal Mode",
+    "sg1_hot_water_and_heating_off": "SG1 - Hot Water and Heating Off",
+    "sg2_hot_water_and_heating_plus_5c": "SG2 - Hot Water and Heating +5°C",
+    "sg3_hot_water_70c": "SG3 - Hot Water 70°C"
+}
