+++ components/daikin_rotex_can/translations/it.py
@@ -1,0 +1,105 @@
+"""
+    This file is used to store translations for the Daikin Rotex CAN component.
+    Language: Italian
+"""
+
+translations_it = {
+    "heating": "Riscaldamento",
+    "cooling": "Raffreddamento",
+    "standby": "Standby",
+    "hot_water_production": "Produzione di Acqua Calda",
+    "defrosting": "Sbrinamento",
+    "optimized_defrosting": "Sbrinamento Ottimizzato",
+    "temperature_antifreeze": "Temperatura Antigelo",
+    "off": "Spento",
+    "on": "Acceso",
+    "night_only": "Solo di Notte",							  
+    "monday": "Lunedì",
+    "tuesday": "Martedì",
+    "wednesday": "Mercoledì",
+    "thursday": "Giovedì",
+    "friday": "Venerdì",
+    "saturday": "Sabato",
+    "sunday": "Domenica",
+    "mo_to_su": "Lunedì a Domenica",
+    "lowering": "Abbassamento",
+    "summer": "Estate",
+    "automatic_1": "Automatico 1",
+    "automatic_2": "Automatico 2",
+    "weather_dependent": "Dipendente dal Tempo",		
+    "standby_mode": "Pronto",
+    "heating_mode": "Riscaldamento",
+    "cooling_mode": "Raffreddamento",
+    "defrosting_mode": "Sbrinamento",
+    "hot_water_production_mode": "Produzione di Acqua Calda",
+    "sg_mode_1": "Modalità SG 1",
+    "sg_mode_2": "Modalità SG 2",
+    "fixed": "Fisso",
+    "no_additional_heat_generator": "Nessun Generatore di Calore Aggiuntivo",
+    "optional_backup_heater": "Backup-Heater Opzionale",
+    "wez_for_hot_water_and_heating": "WEZ per Acqua Calda e Riscaldamento",
+    "wez1_for_hot_water_wez2_for_heating": "WEZ1 per Acqua Calda - WEZ2 per Riscaldamento",
+    "err_0": "Nessun Errore",
+    "err_E9001": "E9001 Sensore di Ritorno",
+    "err_E9002": "E9002 Sensore di Mandata",
+    "err_E9003": "E9003 Funzione Protezione Antigelo",
+    "err_E9004": "E9004 Flusso",
+    "err_E9005": "E9005 Sensore Temperatura Mandata",
+    "err_E9006": "E9006 Sensore Temperatura Mandata",
+    "err_E9007": "E9007 Scheda IG Difettosa",
+    "err_E9008": "E9008 Temperatura Refrigerante Fuori Range",
+    "err_E9009": "E9009 Errore STB",
+    "err_E9010": "E9010 Errore STB",
+    "err_E9011": "E9011 Errore Sensore di Flusso",
+    "err_E9012": "E9012 Errore Sensore Mandata",
+    "err_E9013": "E9013 Scheda AG Difettosa",
+    "err_E9014": "E9014 Pressione Refrigerante Alta",
+    "err_E9015": "E9015 Pressione Refrigerante Bassa",
+    "err_E9016": "E9016 Protezione Sovraccarico Compressore",
+    "err_E9017": "E9017 Ventilatore Bloccato",
+    "err_E9018": "E9018 Valvola di Espansione",
+    "err_E9019": "E9019 Temperatura Acqua Calda > 85°C",
+    "err_E9020": "E9020 Temperatura Evaporatore Alta",
+    "err_E9021": "E9021 Sistema HPS",
+    "err_E9022": "E9022 Errore Sensore Temperatura Ambiente",
+    "err_E9023": "E9023 Errore Sensore Acqua Calda",
+    "err_E9024": "E9024 Sensore di Pressione",
+    "err_E9025": "E9025 Errore Sensore di Ritorno",
+    "err_E9026": "E9026 Sensore di Pressione",
+    "err_E9027": "E9027 Sensore Sbrinamento Air Coil",
+    "err_E9028": "E9028 Sensore Temperatura Air Coil",
+    "err_E9029": "E9029 Errore Sensore Raffreddamento AG",
+    "err_E9030": "E9030 Difetto Elettrico",
+    "err_E9031": "E9031 Difetto Elettrico",
+    "err_E9032": "E9032 Difetto Elettrico",
+    "err_E9033": "E9033 Difetto Elettrico",
+    "err_E9034": "E9034 Difetto Elettrico",
+    "err_E9035": "E9035 Scheda AG Difettosa",
+    "err_E9036": "E9036 Difetto Elettrico",
+    "err_E9037": "E9037 Impostazione Potenza",
+    "err_E9038": "E9038 Perdita di Refrigerante",
+    "err_E9039": "E9039 Sotto/Sovratensione",
+    "err_E9041": "E9041 Errore di Comunicazione",
+    "err_E9042": "E9042 Errore di Comunicazione",
+    "err_E9043": "E9043 Errore di Comunicazione",
+    "err_E9044": "E9044 Errore di Comunicazione",
+    "err_E75": "E75 Errore Sensore Temperatura Ambiente",
+    "err_E76": "E76 Errore Sensore Temperatura Serbatoio",
+    "err_E81": "E81 Errore di Comunicazione Rocon",
+    "err_E88": "E88 Errore di Comunicazione Manuale Rocon",
+    "err_E91": "E91 Errore di Comunicazione Manuale Rocon",
+    "err_E128": "E128 Errore Sensore Temperatura di Ritorno",
+    "err_E129": "E129 Errore Sensore di Pressione",
+    "err_E198": "E198 Misurazione Flusso Non Plausibile",
+    "err_E200": "E200 Errore di Comunicazione",
+    "err_E8005": "E8005 Pressione dell'Acqua nel Sistema di Riscaldamento Troppo Bassa",
+    "err_E8100": "E8100 Comunicazione",
+    "err_E9000": "E9000 Messaggio Interno Temporaneo",
+    "err_W8006": "W8006 Avviso di Perdita di Pressione",
+    "err_W8007": "W8007 Pressione dell'Acqua nel Sistema Troppo Alta",
+    "defect": "Malfunzionamento",
+    "sgn_normal_mode": "SGN - Modalità Normale",
+    "sg1_hot_water_and_heating_off": "SG1 - Acqua Calda e Riscaldamento Spenti",
+    "sg2_hot_water_and_heating_plus_5c": "SG2 - Acqua Calda e Riscaldamento +5°C",
+    "sg3_hot_water_70c": "SG3 - Acqua Calda 70°C"
+}
